requirements.txt
kafka-python==2.0.2
Mastodon.py==1.6.1
python-dotenv==0.19.0
